requirements.txt
grpcio_tools == 1.33.2
grpcio-reflection == 1.33.2
python-dotenv == 1.0.0
Flask-SQLAlchemy == 3.0.3